pyproject.toml
[project]
name = "dwf-analyzer-mcp"
version = "0.1.0"
description = "MCP Server for DWF drawing file analysis using Amazon Bedrock"
authors = [
    {name = "DWF Analysis Team", email = "team@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["mcp", "dwf", "cad", "drawing", "analysis", "bedrock", "nova"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "fastmcp>=2.10.0",
    "boto3>=1.35.0",
    "pillow>=10.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/jesamkim/dwf-analyzer-mcp"
Repository = "https://github.com/jesamkim/dwf-analyzer-mcp"
Documentation = "https://github.com/jesamkim/dwf-analyzer-mcp/blob/main/README.md"
Issues = "https://github.com/jesamkim/dwf-analyzer-mcp/issues"

[project.scripts]
dwf-analyzer-mcp = "dwf_analyzer_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
